Split benchmarks into multiple files.
This helps LLVM optimize, but may be unrealistically favorable,
so we'll need further benchmarks where the same code is used
from multiple so that it won't be a single-caller function.

On the other hand, some usage will be single-caller, so these
benchmarks are useful for those scenarios, and they allow
for fair comparisons between types as various optimizations
are attempted.